+++ src/theme/components/README.md
@@ -1,5 +1,5 @@
 # Shared Components
 
-Some themes may not need to reinvent the wheel by creating brand new components. There are some
-components available to be shared between themes. Any common-ground component that gets shared
-between themes can live here.
+Some themes may not need to create brand new components. Some components are available to share
+between themes. Any common-ground component that gets shared between themes can live here, in the
+`theme/components/` folder.
